feat(theme): implement theme patching for Linux using DBus
